# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AAST Trainery API's",
        "description": "Trainery API documentation",
        "contact": {
            "email": "admin@aast-trainery.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Trainery Demo Server"
        }
    ],
    "paths": {
...
        "/A/student/get-profile": {
            "get": {
                "tags": [
                    "A-Student Profile"
                ],
                "summary": "Get Student Profile ",
                "description": " Get Student Profile ",
                "operationId": "AgetStudentProfile",
                "responses": {
                    "200": {
                        "description": "Student Data to success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessOkVirtual"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response422Virtual"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response401Virtual"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SuccessOkVirtual": {
                "title": "success 200",
                "description": "success 200",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 200
                    },
                    "message": {
                        "title": "message",
                        "description": "Success message",
                        "type": "string",
                        "example": "success"
                    },
                    "response": {
                        "title": "response",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "description": "null or data sent in the request body",
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Response401Virtual": {
                "title": "response Error 401",
                "description": "response Error 401",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 401
                    },
                    "message": {
                        "title": "message",
                        "description": "unauthorized",
                        "type": "string",
                        "example": "unauthorized user"
                    },
                    "response": {
                        "title": "data",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "title": "errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Response422Virtual": {
                "title": "response validation Error 422",
                "description": "response validation Error 422",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "example": 422
                    },
                    "message": {
                        "title": "message",
                        "description": "invalid data",
                        "type": "string",
                        "example": "The given data was invalid"
                    },
                    "response": {
                        "title": "data",
                        "description": "data object",
                        "properties": {
                            "data": {
                                "type": "object"
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The email field is required"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "http",
                "description": "<b>User Token:-<b><br>\n                                        <b>Header key:</b> <i>Authorization</i><br>\n                                        <b>Scheme:</b> <i>bearer</i><br>\n                                        <b>Example:</b> <i>Authorization: Bearer {JWT_token}</i>\n                                        ",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/api/1.0/auth/login",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}